# crates/vault-api/src/handlers/datasets.rs
//! Dataset handlers.

use crate::{
    error::ApiError,
    pagination::{PagedResponse, Pagination},
    response::{ApiResponse, Created, JsonResponse, NoContent},
    state::AppState,
    validation::{validate_dataset_id, Validator},
    ApiResult,
};
use axum::{
    extract::{Path, Query, State},
    Json,
};
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use std::sync::Arc;
use uuid::Uuid;
use validator::Validate;

/// Dataset response.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct DatasetResponse {
    /// Dataset ID.
    pub id: String,
    /// Dataset name.
    pub name: String,
    /// Description.
    pub description: Option<String>,
    /// Schema version.
    pub schema_version: String,
    /// Record count.
    pub record_count: u64,
    /// Total size in bytes.
    pub size_bytes: u64,
    /// Created timestamp.
    pub created_at: DateTime<Utc>,
    /// Updated timestamp.
    pub updated_at: DateTime<Utc>,
    /// Created by user ID.
    pub created_by: String,
    /// Status.
    pub status: DatasetStatus,
    /// Tags.
    pub tags: Vec<String>,
}

/// Dataset status.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
#[serde(rename_all = "lowercase")]
pub enum DatasetStatus {
    /// Active dataset.
    Active,
    /// Archived dataset.
    Archived,
    /// Deleted (soft delete).
    Deleted,
}

/// Create dataset request.
#[derive(Debug, Deserialize, Validate)]
pub struct CreateDatasetRequest {
    /// Dataset name.
    #[validate(length(min = 1, max = 255))]
    pub name: String,
    /// Description.
    #[validate(length(max = 1000))]
    pub description: Option<String>,
    /// Schema (JSON).
    pub schema: Option<serde_json::Value>,
    /// Tags.
    pub tags: Option<Vec<String>>,
}

/// Update dataset request.
#[derive(Debug, Deserialize, Validate)]
pub struct UpdateDatasetRequest {
    /// Dataset name.
    #[validate(length(min = 1, max = 255))]
    pub name: Option<String>,
    /// Description.
    #[validate(length(max = 1000))]
    pub description: Option<String>,
    /// Tags.
    pub tags: Option<Vec<String>>,
}

/// Dataset query parameters.
#[derive(Debug, Deserialize)]
pub struct DatasetQuery {
    /// Filter by status.
    pub status: Option<DatasetStatus>,
    /// Filter by tag.
    pub tag: Option<String>,
    /// Search query.
    pub q: Option<String>,
    /// Pagination.
    #[serde(flatten)]
    pub pagination: Pagination,
}

/// List datasets handler.
pub async fn list_datasets(
    State(state): State<Arc<AppState>>,
    Query(query): Query<DatasetQuery>,
) -> ApiResult<Json<PagedResponse<DatasetResponse>>> {
    // In a real implementation, query the database
    // For now, return empty list
    let response = PagedResponse::empty(&query.pagination);
    Ok(Json(response))
}

/// Get dataset handler.
pub async fn get_dataset(
    State(state): State<Arc<AppState>>,
    Path(id): Path<String>,
) -> ApiResult<Json<DatasetResponse>> {
    validate_dataset_id(&id)?;

    // In a real implementation, fetch from database
    // For now, return not found
    Err(ApiError::not_found_resource("Dataset", &id))
}

/// Create dataset handler.
pub async fn create_dataset(
    State(state): State<Arc<AppState>>,
    Json(request): Json<CreateDatasetRequest>,
) -> ApiResult<Created<DatasetResponse>> {
    // Validate request
    request.validate()?;

    Validator::new()
        .required("name", &request.name)
        .min_length("name", &request.name, 1)
        .finish()?;

    let now = Utc::now();
    let id = Uuid::new_v4().to_string();

    let dataset = DatasetResponse {
        id: id.clone(),
        name: request.name,
        description: request.description,
        schema_version: "1.0".to_string(),
        record_count: 0,
        size_bytes: 0,
        created_at: now,
        updated_at: now,
        created_by: "system".to_string(), // Would come from auth context
        status: DatasetStatus::Active,
        tags: request.tags.unwrap_or_default(),
    };

    // In a real implementation, persist to database

    Ok(Created::new(dataset, format!("/api/v1/datasets/{}", id)))
}

/// Update dataset handler.
pub async fn update_dataset(
    State(state): State<Arc<AppState>>,
    Path(id): Path<String>,
    Json(request): Json<UpdateDatasetRequest>,
) -> ApiResult<Json<DatasetResponse>> {
    validate_dataset_id(&id)?;
    request.validate()?;

    // In a real implementation, update in database
    Err(ApiError::not_found_resource("Dataset", &id))
}

/// Delete dataset handler.
pub async fn delete_dataset(
    State(state): State<Arc<AppState>>,
    Path(id): Path<String>,
) -> ApiResult<NoContent> {
    validate_dataset_id(&id)?;

    // In a real implementation, soft delete in database
    Err(ApiError::not_found_resource("Dataset", &id))
}

/// Archive dataset handler.
pub async fn archive_dataset(
    State(state): State<Arc<AppState>>,
    Path(id): Path<String>,
) -> ApiResult<Json<DatasetResponse>> {
    validate_dataset_id(&id)?;

    // In a real implementation, archive the dataset
    Err(ApiError::not_found_resource("Dataset", &id))
}

/// Restore dataset handler.
pub async fn restore_dataset(
    State(state): State<Arc<AppState>>,
    Path(id): Path<String>,
) -> ApiResult<Json<DatasetResponse>> {
    validate_dataset_id(&id)?;

    // In a real implementation, restore the dataset
    Err(ApiError::not_found_resource("Dataset", &id))
}

/// Dataset statistics response.
#[derive(Debug, Serialize)]
pub struct DatasetStats {
    /// Total records.
    pub total_records: u64,
    /// Total size in bytes.
    pub total_size_bytes: u64,
    /// Unique values by field.
    pub unique_values: serde_json::Value,
    /// Field statistics.
    pub field_stats: serde_json::Value,
}

/// Get dataset statistics handler.
pub async fn get_dataset_stats(
    State(state): State<Arc<AppState>>,
    Path(id): Path<String>,
) -> ApiResult<Json<DatasetStats>> {
    validate_dataset_id(&id)?;

    // In a real implementation, compute statistics
    Err(ApiError::not_found_resource("Dataset", &id))
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_create_dataset_validation() {
        let request = CreateDatasetRequest {
            name: "test".to_string(),
            description: Some("Test dataset".to_string()),
            schema: None,
            tags: None,
        };

        assert!(request.validate().is_ok());
    }

    #[test]
    fn test_create_dataset_empty_name() {
        let request = CreateDatasetRequest {
            name: "".to_string(),
            description: None,
            schema: None,
            tags: None,
        };

        assert!(request.validate().is_err());
    }
}