# crates/vault-api/src/handlers/records.rs
//! Record handlers.

use crate::{
    error::ApiError,
    pagination::{CursorPagedResponse, CursorPagination, PagedResponse, Pagination},
    response::{Created, NoContent},
    state::AppState,
    validation::validate_dataset_id,
    ApiResult,
};
use axum::{
    extract::{Path, Query, State},
    Json,
};
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use std::sync::Arc;
use uuid::Uuid;
use validator::Validate;

/// Record response.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct RecordResponse {
    /// Record ID.
    pub id: String,
    /// Dataset ID.
    pub dataset_id: String,
    /// Record data.
    pub data: serde_json::Value,
    /// Content hash.
    pub content_hash: String,
    /// Version number.
    pub version: u64,
    /// Created timestamp.
    pub created_at: DateTime<Utc>,
    /// Updated timestamp.
    pub updated_at: DateTime<Utc>,
    /// Metadata.
    pub metadata: Option<serde_json::Value>,
}

/// Create record request.
#[derive(Debug, Serialize, Deserialize, Validate)]
pub struct CreateRecordRequest {
    /// Record data.
    pub data: serde_json::Value,
    /// Optional ID (auto-generated if not provided).
    pub id: Option<String>,
    /// Metadata.
    pub metadata: Option<serde_json::Value>,
}

/// Update record request.
#[derive(Debug, Serialize, Deserialize, Validate)]
pub struct UpdateRecordRequest {
    /// Record data.
    pub data: serde_json::Value,
    /// Metadata.
    pub metadata: Option<serde_json::Value>,
    /// Expected version (for optimistic locking).
    pub expected_version: Option<u64>,
}

/// Batch create request.
#[derive(Debug, Deserialize, Validate)]
pub struct BatchCreateRequest {
    /// Records to create.
    #[validate(length(min = 1, max = 1000))]
    pub records: Vec<CreateRecordRequest>,
}

/// Batch operation response.
#[derive(Debug, Serialize)]
pub struct BatchResponse {
    /// Successfully created record IDs.
    pub created: Vec<String>,
    /// Failed records with errors.
    pub failed: Vec<BatchError>,
    /// Total processed.
    pub total: usize,
}

/// Batch error.
#[derive(Debug, Serialize)]
pub struct BatchError {
    /// Index in the batch.
    pub index: usize,
    /// Error message.
    pub error: String,
}

/// Record query parameters.
#[derive(Debug, Deserialize)]
pub struct RecordQuery {
    /// Filter expression.
    pub filter: Option<String>,
    /// Select fields.
    pub select: Option<String>,
    /// Pagination.
    #[serde(flatten)]
    pub pagination: CursorPagination,
}

/// List records handler.
pub async fn list_records(
    State(state): State<Arc<AppState>>,
    Path(dataset_id): Path<String>,
    Query(query): Query<RecordQuery>,
) -> ApiResult<Json<CursorPagedResponse<RecordResponse>>> {
    validate_dataset_id(&dataset_id)?;

    // In a real implementation, query the database
    let response = CursorPagedResponse::empty();
    Ok(Json(response))
}

/// Get record handler.
pub async fn get_record(
    State(state): State<Arc<AppState>>,
    Path((dataset_id, record_id)): Path<(String, String)>,
) -> ApiResult<Json<RecordResponse>> {
    validate_dataset_id(&dataset_id)?;

    // In a real implementation, fetch from database
    Err(ApiError::not_found_resource("Record", &record_id))
}

/// Create record handler.
pub async fn create_record(
    State(state): State<Arc<AppState>>,
    Path(dataset_id): Path<String>,
    Json(request): Json<CreateRecordRequest>,
) -> ApiResult<Created<RecordResponse>> {
    validate_dataset_id(&dataset_id)?;

    let now = Utc::now();
    let id = request.id.unwrap_or_else(|| Uuid::new_v4().to_string());

    // Compute content hash
    let content = serde_json::to_vec(&request.data)?;
    let hash = blake3::hash(&content).to_hex().to_string();

    let record = RecordResponse {
        id: id.clone(),
        dataset_id: dataset_id.clone(),
        data: request.data,
        content_hash: hash,
        version: 1,
        created_at: now,
        updated_at: now,
        metadata: request.metadata,
    };

    // In a real implementation, persist to database

    Ok(Created::new(
        record,
        format!("/api/v1/datasets/{}/records/{}", dataset_id, id),
    ))
}

/// Update record handler.
pub async fn update_record(
    State(state): State<Arc<AppState>>,
    Path((dataset_id, record_id)): Path<(String, String)>,
    Json(request): Json<UpdateRecordRequest>,
) -> ApiResult<Json<RecordResponse>> {
    validate_dataset_id(&dataset_id)?;

    // In a real implementation, update in database with optimistic locking
    Err(ApiError::not_found_resource("Record", &record_id))
}

/// Delete record handler.
pub async fn delete_record(
    State(state): State<Arc<AppState>>,
    Path((dataset_id, record_id)): Path<(String, String)>,
) -> ApiResult<NoContent> {
    validate_dataset_id(&dataset_id)?;

    // In a real implementation, soft delete in database
    Err(ApiError::not_found_resource("Record", &record_id))
}

/// Batch create records handler.
pub async fn batch_create_records(
    State(state): State<Arc<AppState>>,
    Path(dataset_id): Path<String>,
    Json(request): Json<BatchCreateRequest>,
) -> ApiResult<Json<BatchResponse>> {
    validate_dataset_id(&dataset_id)?;
    request.validate()?;

    let mut created = Vec::new();
    let mut failed = Vec::new();
    let total = request.records.len();

    for (index, record_request) in request.records.into_iter().enumerate() {
        // In a real implementation, create each record
        // For now, simulate success
        let id = record_request.id.unwrap_or_else(|| Uuid::new_v4().to_string());
        created.push(id);
    }

    Ok(Json(BatchResponse {
        created,
        failed,
        total,
    }))
}

/// Batch delete request.
#[derive(Debug, Deserialize, Validate)]
pub struct BatchDeleteRequest {
    /// Record IDs to delete.
    #[validate(length(min = 1, max = 1000))]
    pub ids: Vec<String>,
}

/// Batch delete records handler.
pub async fn batch_delete_records(
    State(state): State<Arc<AppState>>,
    Path(dataset_id): Path<String>,
    Json(request): Json<BatchDeleteRequest>,
) -> ApiResult<Json<BatchResponse>> {
    validate_dataset_id(&dataset_id)?;
    request.validate()?;

    let total = request.ids.len();

    // In a real implementation, delete records
    Ok(Json(BatchResponse {
        created: vec![], // Use for deleted IDs
        failed: vec![],
        total,
    }))
}

/// Record history entry.
#[derive(Debug, Serialize)]
pub struct RecordHistoryEntry {
    /// Version number.
    pub version: u64,
    /// Content hash.
    pub content_hash: String,
    /// Timestamp.
    pub timestamp: DateTime<Utc>,
    /// User who made the change.
    pub user_id: String,
    /// Change type.
    pub change_type: String,
}

/// Get record history handler.
pub async fn get_record_history(
    State(state): State<Arc<AppState>>,
    Path((dataset_id, record_id)): Path<(String, String)>,
    Query(pagination): Query<Pagination>,
) -> ApiResult<Json<PagedResponse<RecordHistoryEntry>>> {
    validate_dataset_id(&dataset_id)?;

    // In a real implementation, fetch history from database
    Err(ApiError::not_found_resource("Record", &record_id))
}

/// Get specific record version handler.
pub async fn get_record_version(
    State(state): State<Arc<AppState>>,
    Path((dataset_id, record_id, version)): Path<(String, String, u64)>,
) -> ApiResult<Json<RecordResponse>> {
    validate_dataset_id(&dataset_id)?;

    // In a real implementation, fetch specific version
    Err(ApiError::not_found_resource("Record", &record_id))
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_create_record_request() {
        let request = CreateRecordRequest {
            data: serde_json::json!({"name": "test"}),
            id: None,
            metadata: None,
        };

        // Validation should pass
        assert!(request.validate().is_ok());
    }

    #[test]
    fn test_batch_create_validation() {
        let request = BatchCreateRequest {
            records: vec![CreateRecordRequest {
                data: serde_json::json!({"field": "value"}),
                id: None,
                metadata: None,
            }],
        };

        assert!(request.validate().is_ok());
    }

    #[test]
    fn test_batch_create_empty() {
        let request = BatchCreateRequest { records: vec![] };

        assert!(request.validate().is_err());
    }
}