# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Data Vault API
  description: |
    Enterprise-grade secure data management API for LLM training pipelines.

    ## Authentication

    Most endpoints require authentication via JWT bearer token or API key.

    ### JWT Authentication
    1. Call `/api/v1/auth/login` with credentials
    2. Use the returned `access_token` in the `Authorization` header
    3. Refresh tokens using `/api/v1/auth/refresh` before expiration

    ### API Key Authentication
    Include your API key in the `X-API-Key` header.

    ## Rate Limiting

    - Default: 100 requests/second per client
    - Burst: 200 requests
    - Rate limit headers included in responses

    ## Pagination

    List endpoints support cursor-based pagination:
    - `limit`: Number of items (default: 20, max: 100)
    - `cursor`: Cursor from previous response

    ## Error Responses

    All errors return JSON with:
    - `code`: Machine-readable error code
    - `message`: Human-readable description
    - `request_id`: For support inquiries

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
    url: https://docs.example.com
  license:
    name: LLMDevOps-PSACL v1.0
    url: https://example.com/license
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local Development
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/datasets/{datasetId}:
    parameters:
      - $ref: '#/components/parameters/DatasetId'

    get:
      tags: [Datasets]
      summary: Get dataset
      description: Returns details of a specific dataset.
      operationId: getDataset
      responses:
        '200':
          description: Dataset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Datasets]
      summary: Update dataset
      description: Updates dataset metadata.
      operationId: updateDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDatasetRequest'
      responses:
        '200':
          description: Dataset updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Datasets]
      summary: Delete dataset
      description: Soft-deletes a dataset. Can be restored within retention period.
      operationId: deleteDataset
      responses:
        '204':
          description: Dataset deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/datasets/{datasetId}/records/{recordId}:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
      - $ref: '#/components/parameters/RecordId'

    get:
      tags: [Records]
      summary: Get record
      description: Returns a specific record.
      operationId: getRecord
      responses:
        '200':
          description: Record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Records]
      summary: Update record
      description: Updates a record. Creates a new version.
      operationId: updateRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecordRequest'
      responses:
        '200':
          description: Record updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Records]
      summary: Delete record
      description: Soft-deletes a record.
      operationId: deleteRecord
      responses:
        '204':
          description: Record deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  parameters:
    DatasetId:
      name: datasetId
      in: path
      required: true
      description: Dataset unique identifier
      schema:
        type: string
        format: uuid
    RecordId:
      name: recordId
      in: path
      required: true
      description: Record unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired token
            request_id: req_abc123
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Dataset with id 'ds_123' not found
            request_id: req_abc123
    Conflict:
      description: Resource conflict (e.g., already exists)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: Resource already exists
            request_id: req_abc123
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            errors:
              email:
                - must be a valid email address
              password:
                - must be at least 8 characters
            request_id: req_abc123
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Request ID for support
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    UpdateDatasetRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        tags:
          type: array
          items:
            type: string
          maxItems: 20
        settings:
          type: object
          additionalProperties: true
    DatasetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        record_count:
          type: integer
        size_bytes:
          type: integer
          format: int64
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateRecordRequest:
      type: object
      required: [data]
      properties:
        data:
          type: object
        metadata:
          type: object
        expected_version:
          type: integer
          description: For optimistic locking
    RecordResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        data:
          type: object
        content_hash:
          type: string
        version:
          type: integer
        pii_detections:
          type: array
          items:
            $ref: '#/components/schemas/PIIDetection'
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PIIDetection:
      type: object
      properties:
        type:
          type: string
          description: Type of PII detected
        value:
          type: string
          description: The detected value (may be masked)
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        risk_level:
          type: string
          enum: [low, medium, high, critical]
        location:
          type: object
          properties:
            start:
              type: integer
            end:
              type: integer
...